added base test parent class for cleaner code
